+++ src/trim_slice.rs
@@ -6,6 +6,7 @@ use alloc::{
 	boxed::Box,
 	vec::Vec,
 };
+use crate::pattern::MatchPattern;
 
 
 
@@ -19,51 +20,75 @@ use alloc::{
 ///
 /// | Method | Description |
 /// | ------ | ----------- |
-/// | `trim_matches` | Trim arbitrary leading and trailing bytes via callback. |
-/// | `trim_start_matches` | Trim arbitrary leading bytes via callback. |
-/// | `trim_end_matches` | Trim arbitrary trailing bytes via callback. |
+/// | `trim_matches` | Trim arbitrary leading and trailing bytes. |
+/// | `trim_start_matches` | Trim arbitrary leading bytes. |
+/// | `trim_end_matches` | Trim arbitrary trailing bytes. |
+///
+/// Each of these match methods accept either:
+/// * A single `u8`;
+/// * An array or slice of `u8`;
+/// * A `&BTreeSet<u8>`;
+/// * A callback with the signature `Fn(u8) -> bool`;
 pub trait TrimSliceMatches {
 	/// # Trim Matches.
 	///
 	/// Trim arbitrary leading and trailing bytes as determined by the provided
-	/// callback, where a return value of `true` means trim.
+	/// pattern, which can be:
+	/// * A single `u8`;
+	/// * An array or slice of `u8`;
+	/// * A `&BTreeSet<u8>`;
+	/// * A callback with the signature `Fn(u8) -> bool`;
 	///
 	/// ```
 	/// use trimothy::TrimSliceMatches;
 	///
 	/// let s: &[u8] = b"...Custom Trim!...";
+	/// assert_eq!(s.trim_matches(b'.'), b"Custom Trim!");
+	/// assert_eq!(s.trim_matches([b'.']), b"Custom Trim!");
+	/// assert_eq!(s.trim_matches(&[b'.']), b"Custom Trim!");
 	/// assert_eq!(s.trim_matches(|b| b'.' == b), b"Custom Trim!");
 	/// ```
-	fn trim_matches<F>(&self, cb: F) -> &[u8]
-	where F: Fn(u8) -> bool;
+	fn trim_matches<P: MatchPattern<u8>>(&self, pat: P) -> &[u8];
 
 	/// # Trim Start Matches.
 	///
-	/// Trim arbitrary leading bytes as determined by the provided callback,
-	/// where a return value of `true` means trim.
+	/// Trim arbitrary leading bytes as determined by the provided
+	/// pattern, which can be:
+	/// * A single `u8`;
+	/// * An array or slice of `u8`;
+	/// * A `&BTreeSet<u8>`;
+	/// * A callback with the signature `Fn(u8) -> bool`;
 	///
 	/// ```
 	/// use trimothy::TrimSliceMatches;
 	///
 	/// let s: &[u8] = b"...Custom Trim!...";
+	/// assert_eq!(s.trim_start_matches(b'.'), b"Custom Trim!...");
+	/// assert_eq!(s.trim_start_matches([b'.']), b"Custom Trim!...");
+	/// assert_eq!(s.trim_start_matches(&[b'.']), b"Custom Trim!...");
 	/// assert_eq!(s.trim_start_matches(|b| b'.' == b), b"Custom Trim!...");
 	/// ```
-	fn trim_start_matches<F>(&self, cb: F) -> &[u8]
-	where F: Fn(u8) -> bool;
+	fn trim_start_matches<P: MatchPattern<u8>>(&self, pat: P) -> &[u8];
 
 	/// # Trim End Matches.
 	///
-	/// Trim arbitrary trailing bytes as determined by the provided callback,
-	/// where a return value of `true` means trim.
+	/// Trim arbitrary trailing bytes as determined by the provided
+	/// pattern, which can be:
+	/// * A single `u8`;
+	/// * An array or slice of `u8`;
+	/// * A `&BTreeSet<u8>`;
+	/// * A callback with the signature `Fn(u8) -> bool`;
 	///
 	/// ```
 	/// use trimothy::TrimSliceMatches;
 	///
 	/// let s: &[u8] = b"...Custom Trim!...";
+	/// assert_eq!(s.trim_end_matches(b'.'), b"...Custom Trim!");
+	/// assert_eq!(s.trim_end_matches([b'.']), b"...Custom Trim!");
+	/// assert_eq!(s.trim_end_matches(&[b'.']), b"...Custom Trim!");
 	/// assert_eq!(s.trim_end_matches(|b| b'.' == b), b"...Custom Trim!");
 	/// ```
-	fn trim_end_matches<F>(&self, cb: F) -> &[u8]
-	where F: Fn(u8) -> bool;
+	fn trim_end_matches<P: MatchPattern<u8>>(&self, pat: P) -> &[u8];
 }
 
 
@@ -74,17 +99,20 @@ macro_rules! trim_slice {
 			/// # Trim Matches.
 			///
 			/// Trim arbitrary leading and trailing bytes as determined by the provided
-			/// callback, where a return value of `true` means trim.
-			fn trim_matches<F>(&self, cb: F) -> &[u8]
-			where F: Fn(u8) -> bool {
+			/// pattern, which can be:
+			/// * A single `u8`;
+			/// * An array or slice of `u8`;
+			/// * A `&BTreeSet<u8>`;
+			/// * A callback with the signature `Fn(u8) -> bool`;
+			fn trim_matches<P: MatchPattern<u8>>(&self, pat: P) -> &[u8] {
 				let mut src: &[u8] = &self;
 				while let [first, rest @ ..] = src {
-					if cb(*first) { src = rest; }
+					if pat.is_match(*first) { src = rest; }
 					else { break; }
 				}
 
 				while let [rest @ .., last] = src {
-					if cb(*last) { src = rest; }
+					if pat.is_match(*last) { src = rest; }
 					else { break; }
 				}
 				src
@@ -92,13 +120,16 @@ macro_rules! trim_slice {
 
 			/// # Trim Start Matches.
 			///
-			/// Trim arbitrary leading bytes as determined by the provided callback,
-			/// where a return value of `true` means trim.
-			fn trim_start_matches<F>(&self, cb: F) -> &[u8]
-			where F: Fn(u8) -> bool {
+			/// Trim arbitrary leading bytes as determined by the provided
+			/// pattern, which can be:
+			/// * A single `u8`;
+			/// * An array or slice of `u8`;
+			/// * A `&BTreeSet<u8>`;
+			/// * A callback with the signature `Fn(u8) -> bool`;
+			fn trim_start_matches<P: MatchPattern<u8>>(&self, pat: P) -> &[u8] {
 				let mut src: &[u8] = &self;
 				while let [first, rest @ ..] = src {
-					if cb(*first) { src = rest; }
+					if pat.is_match(*first) { src = rest; }
 					else { break; }
 				}
 				src
@@ -106,13 +137,16 @@ macro_rules! trim_slice {
 
 			/// # Trim Start Matches.
 			///
-			/// Trim arbitrary leading bytes as determined by the provided callback,
-			/// where a return value of `true` means trim.
-			fn trim_end_matches<F>(&self, cb: F) -> &[u8]
-			where F: Fn(u8) -> bool {
+			/// Trim arbitrary trailing bytes as determined by the provided
+			/// pattern, which can be:
+			/// * A single `u8`;
+			/// * An array or slice of `u8`;
+			/// * A `&BTreeSet<u8>`;
+			/// * A callback with the signature `Fn(u8) -> bool`;
+			fn trim_end_matches<P: MatchPattern<u8>>(&self, pat: P) -> &[u8] {
 				let mut src: &[u8] = &self;
 				while let [rest @ .., last] = src {
-					if cb(*last) { src = rest; }
+					if pat.is_match(*last) { src = rest; }
 					else { break; }
 				}
 				src
@@ -157,17 +191,21 @@ mod tests {
 			assert_eq!(a.trim_ascii(), b);
 		}
 
-		assert_eq!(T_EMPTY.trim_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
-		assert_eq!(T_EMPTY.to_vec().trim_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
-		assert_eq!(Box::<[u8]>::from(T_EMPTY).trim_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(T_EMPTY.trim_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(T_EMPTY.to_vec().trim_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(Box::<[u8]>::from(T_EMPTY).trim_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
 
-		assert_eq!(b"  ".trim_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
-		assert_eq!(b"  ".to_vec().trim_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
-		assert_eq!(Box::<[u8]>::from("  ".as_bytes()).trim_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(b"  ".trim_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(b"  ".to_vec().trim_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(Box::<[u8]>::from("  ".as_bytes()).trim_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
 
-		assert_eq!(T_HELLO_E.trim_matches(|b| b'h' == b), b"ello\t");
-		assert_eq!(T_HELLO_E.to_vec().trim_matches(|b| b'h' == b), b"ello\t");
-		assert_eq!(Box::<[u8]>::from(T_HELLO_E).trim_matches(|b| b'h' == b), b"ello\t");
+		assert_eq!(T_HELLO_E.trim_matches(|b: u8| b'h' == b), b"ello\t");
+		assert_eq!(T_HELLO_E.to_vec().trim_matches(|b: u8| b'h' == b), b"ello\t");
+		assert_eq!(Box::<[u8]>::from(T_HELLO_E).trim_matches(|b: u8| b'h' == b), b"ello\t");
+
+		assert_eq!(T_HELLO_E.trim_matches(b'h'), b"ello\t");
+		assert_eq!(T_HELLO_E.to_vec().trim_matches(b'h'), b"ello\t");
+		assert_eq!(Box::<[u8]>::from(T_HELLO_E).trim_matches(b'h'), b"ello\t");
 
 		// This should also work on arrays.
 		let arr: [u8; 5] = [b' ', b' ', b'.', b' ', b' '];
@@ -197,17 +235,21 @@ mod tests {
 			assert_eq!(a.trim_ascii_start(), b);
 		}
 
-		assert_eq!(T_EMPTY.trim_start_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
-		assert_eq!(T_EMPTY.to_vec().trim_start_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
-		assert_eq!(Box::<[u8]>::from(T_EMPTY).trim_start_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(T_EMPTY.trim_start_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(T_EMPTY.to_vec().trim_start_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(Box::<[u8]>::from(T_EMPTY).trim_start_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
 
-		assert_eq!(b"  ".trim_start_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
-		assert_eq!(b"  ".to_vec().trim_start_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
-		assert_eq!(Box::<[u8]>::from("  ".as_bytes()).trim_start_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(b"  ".trim_start_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(b"  ".to_vec().trim_start_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(Box::<[u8]>::from("  ".as_bytes()).trim_start_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
 
-		assert_eq!(T_HELLO_E.trim_start_matches(|b| b'h' == b), b"ello\t");
-		assert_eq!(Box::<[u8]>::from(T_HELLO_E).trim_start_matches(|b| b'h' == b), b"ello\t");
-		assert_eq!(T_HELLO_E.to_vec().trim_start_matches(|b| b'h' == b), b"ello\t");
+		assert_eq!(T_HELLO_E.trim_start_matches(|b: u8| b'h' == b), b"ello\t");
+		assert_eq!(Box::<[u8]>::from(T_HELLO_E).trim_start_matches(|b: u8| b'h' == b), b"ello\t");
+		assert_eq!(T_HELLO_E.to_vec().trim_start_matches(|b: u8| b'h' == b), b"ello\t");
+
+		assert_eq!(T_HELLO_E.trim_start_matches(b'h'), b"ello\t");
+		assert_eq!(Box::<[u8]>::from(T_HELLO_E).trim_start_matches(b'h'), b"ello\t");
+		assert_eq!(T_HELLO_E.to_vec().trim_start_matches(b'h'), b"ello\t");
 	}
 
 	#[test]
@@ -233,16 +275,20 @@ mod tests {
 			assert_eq!(a.trim_ascii_end(), b);
 		}
 
-		assert_eq!(T_EMPTY.trim_end_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
-		assert_eq!(T_EMPTY.to_vec().trim_end_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
-		assert_eq!(Box::<[u8]>::from(T_EMPTY).trim_end_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(T_EMPTY.trim_end_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(T_EMPTY.to_vec().trim_end_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(Box::<[u8]>::from(T_EMPTY).trim_end_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
 
-		assert_eq!(b"  ".trim_end_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
-		assert_eq!(b"  ".to_vec().trim_end_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
-		assert_eq!(Box::<[u8]>::from("  ".as_bytes()).trim_end_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(b"  ".trim_end_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(b"  ".to_vec().trim_end_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(Box::<[u8]>::from("  ".as_bytes()).trim_end_matches(|b: u8| b.is_ascii_whitespace()), T_EMPTY);
 
-		assert_eq!(T_HELLO_E.trim_matches(|b| b'\t' == b), T_HELLO);
-		assert_eq!(Box::<[u8]>::from(T_HELLO_E).trim_matches(|b| b'\t' == b), T_HELLO);
-		assert_eq!(T_HELLO_E.to_vec().trim_matches(|b| b'\t' == b), T_HELLO);
+		assert_eq!(T_HELLO_E.trim_matches(|b: u8| b'\t' == b), T_HELLO);
+		assert_eq!(Box::<[u8]>::from(T_HELLO_E).trim_matches(|b: u8| b'\t' == b), T_HELLO);
+		assert_eq!(T_HELLO_E.to_vec().trim_matches(|b: u8| b'\t' == b), T_HELLO);
+
+		assert_eq!(T_HELLO_E.trim_matches(b'\t'), T_HELLO);
+		assert_eq!(Box::<[u8]>::from(T_HELLO_E).trim_matches(b'\t'), T_HELLO);
+		assert_eq!(T_HELLO_E.to_vec().trim_matches(b'\t'), T_HELLO);
 	}
 }